builders created and wordleGameState immutable
This change is supposed to make wordleGameState immutable with wasn't
because Tile had accessible variable.Test are also modify in order not
to crash because I changed the naming of the fonction. The is now a
static function startGame.

Hopefully, this will make the code more robust.

This turn out to be really complicated because I think java alows
private inner class access but Koltin doesn't so I had to rewrite every
access.