rules are added and toasts error messages too
The rules of the game should be added. Some message should be displayed
telling if you need to enter a word. This is made using a companion map
in WordleGameActivity so you can decide the game in thi class hovewer
this make WordleMenu depends on the later class which make the code much
more complicated to understand.
Toasts should be added too while playing the game.