GameOptionsMenuActivity commented and changed
One of the functions worry me, I think it could run for a really long
time if the list is too long(I am not certain) anyway I rewritten it
in a more concise way. I also commented the code because I had a hard
time understanding it.

Hopefully, this change will make the code more efficient and clearer

I am worried about having two different event listener in two differents
function. What happen is the list is changed before the size is changed
and the list is Shorter. Would it be possible? that it happens, would the
game crash?