recorder and play use firebase storage
The recording now sends the audio file cached after to the database.
The player now download the audio file from database, cache it and plays
id.

This change aims to be able to sends voicenotes laters if we keep a
general chat that allows to find friends and decide who host the party
and who join the party.